CORS fiddle to mitigate Access-Control-Allow-Origin
Access-Control-Allow-Origin is a problem for an RSS program I am trying to write.  I can achieve this with PHP using 	$xmlDoc = new DOMDocument(); $xmlDoc->load($xml); However the objective is to keep the entire program client side.  2 reasons.
(1) Some feeds are known to limit requests via IP address
(2) In the event that the program goes in to very high usage I would not want local server to carry the burden.